+++ db/migrations/postgresql/2018-04-27-192500_ops_mail/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE mail_domains (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   name VARCHAR(128) NOT NULL,
   created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT now(),
   updated_at TIMESTAMP WITHOUT TIME ZONE NOT NULL
@@ -8,8 +8,8 @@ CREATE TABLE mail_domains (
 CREATE UNIQUE INDEX idx_mail_domains_name ON mail_domains (name);
 
 CREATE TABLE mail_users (
-  id BIGSERIAL PRIMARY KEY,
-  domain_id BIGINT NOT NULL,
+  id SERIAL PRIMARY KEY,
+  domain_id INT NOT NULL,
   email VARCHAR(255) NOT NULL,
   full_name VARCHAR(128) NOT NULL,
   password VARCHAR(255) NOT NULL,
@@ -24,8 +24,8 @@ CREATE UNIQUE INDEX idx_mail_users_email ON mail_users (email);
 CREATE INDEX idx_mail_users_full_name ON mail_users (full_name);
 
 CREATE TABLE mail_aliases (
-  id BIGSERIAL PRIMARY KEY,
-  domain_id BIGINT NOT NULL,
+  id SERIAL PRIMARY KEY,
+  domain_id INT NOT NULL,
   source VARCHAR(255) NOT NULL,
   destination VARCHAR(255) NOT NULL,
   created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT now(),
